Fix: Adjust methods for displaying current user sessions trip from application controller.
