+++ app/controllers/application_controller.rb
@@ -3,5 +3,25 @@ class ApplicationController < ActionController::Base
   include ::ActionController::Cookies
 
   skip_before_action :verify_authenticity_token
-  include sessions_helper
+  helper_method :login!, :logged_in?, :current_user, :authorized_user?, :logout!
+
+  def login!
+    session[:user_id] = @user.id
+  end
+
+  def logged_in?
+    session[:user_id]
+  end
+
+  def current_user
+    return @current_user ||= User.find(session[:user_id]) if session[:user_id]
+  end
+
+  def authorized_user?
+    @user == current_user
+  end
+
+  def logout!
+    session.clear
+  end
 end
