+++ config/routes.rb
@@ -6,6 +6,6 @@ Rails.application.routes.draw do
   delete :logout, to: 'sessions#logout'
   get :logged_in, to: 'sessions#logged_in'
   get :index, to: 'destinies#index'
-  get :list, to: 'trips#index'
+  get :list_trips, to: 'trips#index'
   root to: 'static#home'
 end
